# service/api/get-stream.go
package api

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/julienschmidt/httprouter"
	"github.com/mouvzee/wasaphoto/service/api/methods"
	"github.com/mouvzee/wasaphoto/service/api/reqcontext"
)

/*
getMyStream is the handler for the GET /users/:profileUserID/feed endpoint
It returns the stream of the user with the given profileUserID
*/
func (rt *_router) getMyStream(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
	// Get the profileUserID from the URL
	profileUserID, err := strconv.Atoi(ps.ByName("userID"))
	if err != nil {
		http.Error(w, "Bad Request", http.StatusBadRequest)
		return
	}

	userID := ctx.UserID
	// Check if the user is authorized
	if profileUserID != userID {
		http.Error(w, "Forbidden", http.StatusForbidden)
		return
	}

	// Get limit and offset from the queries
	limit, offset, err := methods.GetLimitAndOffset(r.URL.Query())
	if err != nil {
		http.Error(w, "Bad Request", http.StatusBadRequest)
		return
	}

	// Get the stream
	dbStream, err := rt.db.GetStream(profileUserID, offset, limit)
	if err != nil {
		ctx.Logger.WithError(err).Error("Error getting the stream")
		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		return
	}

	posts := make([]Photo, len(dbStream))

	for i, dbPost := range dbStream {
		var post Photo
		err = post.TakingPhoto(dbPost)
		if err != nil {
			ctx.Logger.WithError(err).Error("Error converting post")
			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
			return
		}
		posts[i] = post
	}

	// Write the response
	w.WriteHeader(http.StatusOK)
	w.Header().Set("Content-Type", "application/json")
	if err := json.NewEncoder(w).Encode(posts); err != nil {
		ctx.Logger.WithError(err).Error("Error encoding the stream")
		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		return
	}

}